feat: moved files in proper folders
Moved the proper framework JS and CSS in folders according to the names.
Moved all the JS functions in src/js
Moved all the PHP functions in src/php

The purpose of moving all the files is for security reasons and it will be easier to navigate different functions.

Refactored add_customer.php, edit.php, and delete.php  especially the queries.
The purpose of refactoring it is for security purposes. Especially for SQL Injection Attacks.