refactor: change home address
Changed home address to 2 different data names (home_address_street for
the Street No./Building No./Barangay field and home_address_city for the
City in the Home Address)
Changed the variable of home address to 2 different data names on php
files.
Changed the way on how the new data names will be displayed for the home
address in the table.